feat: Add a function to get data from GeoNames API
This is the first step of a more complex function that will be
developed. The purpose of this call to GeoNames API is to collect the
geographic coordinates related to the location that the user entered in
the form input.
In this regard, it was necessary to include the Babel transform-runtime
plugin to enable Jest tests on functions that make API calls.